% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_bds}
\alias{fit_bds}
\title{fits the bds model}
\usage{
fit_bds(
  phy,
  sampling_fraction,
  lambda_hat,
  mu_hat,
  eta,
  num_classes = 6,
  sd = 0.587,
  tol = 1e-06
)
}
\arguments{
\item{phy}{an object of type phylo}

\item{sampling_fraction}{the probability that each species was sampled in the tree}

\item{lambda_hat}{the overall scale of the log-normal base distribution for the speciation rates. If not specified, the function will estimate it using ML}

\item{mu_hat}{the overall scale of the log-normal base distribution for the extinction rates. If not specified, the function will estimate it using ML}

\item{eta}{the shift rate parameter. If not specified, the function will estimate it using ML}

\item{num_classes}{the number of rate class discretizations (n), such that rate categories is k = n^2}

\item{sd}{the spread parameter for the log-normal base distribution}

\item{tol}{the local error threshold in the numerical ODE solver (per delta_t time step)}
}
\description{
fits the bds model
}
