% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_cbd}
\alias{fit_cbd}
\title{fits the constant-rate birth-death model}
\usage{
fit_cbd(
  phy,
  sampling_fraction,
  tol = 1e-06,
  condition_survival = TRUE,
  condition_root_speciation = TRUE,
  extinction_approximation = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{phy}{an object of type phylo}

\item{sampling_fraction}{the probability that each species was sampled in the tree}

\item{tol}{the local error threshold in the numerical ODE solver (per delta_t time step)}

\item{condition_survival}{whether or not to condition on the survival of the left and right lineages descending from the root (default TRUE)}

\item{condition_root_speciation}{whether or not to condition on that there was a speciation event at the root node (default TRUE)}
}
\description{
fits the constant-rate birth-death model
}
