% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{Phylogeny}
\alias{Phylogeny}
\alias{$.Phylogeny}
\title{Phylogeny constructor}
\format{
An object of class \code{environment} of length 10.
}
\usage{
Phylogeny(newick)
}
\arguments{
\item{newick}{a newick string}
}
\description{
Constructs a phylogeny that lives in the "rust" part of the program
}
\examples{
Phylogeny$new("((A:0.5,B:0.5):0.5,C:1.0);")

}
\keyword{datasets}
