% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{primates}
\alias{primates}
\title{Primates phylogenetic tree}
\format{
An object of class \code{phylo} of length 5.
}
\usage{
data(primates)
}
\description{
This tree is from Vos & Mooers (2006)
}
\details{
Vos, R. and A. Mooers. 2006. A new dated supertree of the primates. in Inferring large phylogenies: the big tree problem (R Vos, Phd thesis). Simon Fraser University, Burnaby, British Columbia.
}
\keyword{datasets}
