% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPesto.R
\docType{package}
\name{RPesto-package}
\alias{RPesto}
\alias{RPesto-package}
\title{RPesto: Phylogenetic Estimation of Shifts in the Tempo of Origination (in R)}
\description{
This package can estimate branch-specific diversification rates fast and efficiently.
}
\section{References}{

\itemize{
\item Kopperud, B. T., & Höhna, S. (2025). Phylogenetic Estimation of branch-specific Shifts in the Tempo of Origination. Systematic Biology, syaf041.

\item Höhna, S., Freyman, W. A., Nolen, Z., Huelsenbeck, J. P., May, M. R., & Moore, B. R. (2019). A Bayesian approach for estimating branch-specific speciation and extinction rates. BioRxiv, 555805.
}
}

\author{
\strong{Maintainer}: Bjørn Tore Kopperud \email{kopperud@protonmail.com}

Authors:
\itemize{
  \item Sebastian Höhna
}

}
